# [BUGFIX] FastRegexpMatcher: do Unicode normalization as part of case-insensitive comparison