--- a/model/labels/regexp.go
+++ b/model/labels/regexp.go
@@ -16,6 +16,7 @@ package labels
 import (
 	"slices"
 	"strings"
+	"unicode"
 	"unicode/utf8"
 
 	"github.com/grafana/regexp"
@@ -767,7 +768,7 @@ type equalMultiStringMapMatcher struct {
 
 func (m *equalMultiStringMapMatcher) add(s string) {
 	if !m.caseSensitive {
-		s = strings.ToLower(norm.NFKD.String(s))
+		s = toNormalisedLower(s)
 	}
 
 	m.values[s] = struct{}{}
@@ -787,11 +788,48 @@ func (m *equalMultiStringMapMatcher) setMatches() []string {
 
 func (m *equalMultiStringMapMatcher) Matches(s string) bool {
 	if !m.caseSensitive {
-		s = strings.ToLower(norm.NFKD.String(s))
+		s = toNormalisedLower(s)
 	}
 
 	_, ok := m.values[s]
 	return ok
+}
+
+// toNormalisedLower normalise the input string using "Unicode Normalization Form D" and then convert
+// it to lower case.
+func toNormalisedLower(s string) string {
+	// Check if the string is all ASCII chars and convert any upper case character to lower case character.
+	isASCII := true
+	var (
+		b   strings.Builder
+		pos int
+	)
+	b.Grow(len(s))
+	for i := 0; i < len(s); i++ {
+		c := s[i]
+		if isASCII && c >= utf8.RuneSelf {
+			isASCII = false
+		}
+		if 'A' <= c && c <= 'Z' {
+			c += 'a' - 'A'
+			if pos < i {
+				b.WriteString(s[pos:i])
+			}
+			b.WriteByte(c)
+			pos = i + 1
+		}
+	}
+	if pos < len(s) {
+		b.WriteString(s[pos:])
+	}
+
+	// Optimize for ASCII-only strings. In this case we don't have to do any normalization.
+	if isASCII {
+		return b.String()
+	}
+
+	// Normalise and convert to lower.
+	return strings.Map(unicode.ToLower, norm.NFKD.String(b.String()))
 }
 
 // anyStringWithoutNewlineMatcher is a stringMatcher which matches any string