--- a/model/labels/regexp_test.go
+++ b/model/labels/regexp_test.go
@@ -81,6 +81,7 @@ var (
 		".*foo.?",
 		".?foo.+",
 		"foo.?|bar",
+		"ſſs",
 		// Concat of literals and wildcards.
 		".*-.*-.*-.*-.*",
 		"(.+)-(.+)-(.+)-(.+)-(.+)",
@@ -91,7 +92,7 @@ var (
 		"foo", " foo bar", "bar", "buzz\nbar", "bar foo", "bfoo", "\n", "\nfoo", "foo\n", "hello foo world", "hello foo\n world", "",
 		"FOO", "Foo", "fOo", "foO", "OO", "Oo", "\nfoo\n", strings.Repeat("f", 20), "prometheus", "prometheus_api_v1", "prometheus_api_v1_foo",
 		"10.0.1.20", "10.0.2.10", "10.0.3.30", "10.0.4.40",
-		"foofoo0", "foofoo", "😀foo0", "ſſs",
+		"foofoo0", "foofoo", "😀foo0", "ſſs", "ſſS",
 
 		// Values matching / not matching the test regexps on long alternations.
 		"zQPbMkNO", "zQPbMkNo", "jyyfj00j0061", "jyyfj00j006", "jyyfj00j00612", "NNSPdvMi", "NNSPdvMiXXX", "NNSPdvMixxx", "nnSPdvMi", "nnSPdvMiXXX",