@colega 2024-05-31
The test case was defined in [the issue](https://github.com/prometheus/prometheus/issues/14066).