@pracucci 2024-05-31
I've run few selected benchmarks, and there's quite an impactful performance impact:

```
                                                     │ before.txt  │             after.txt              │
                                                     │   sec/op    │   sec/op     vs base               │
FastRegexMatcher/(?i:foo)-12                           115.5n ± 2%   111.7n ± 2%   -3.33% (p=0.004 n=6)
FastRegexMatcher/(?i:(foo|bar))-12                     239.8n ± 4%   233.4n ± 8%        ~ (p=0.288 n=6)
FastRegexMatcher/(?i:(foo1|foo2|bar))-12               431.2n ± 3%   416.6n ± 2%   -3.37% (p=0.002 n=6)
FastRegexMatcher/^(?i:foo|oo)|(bar)$-12                1.050µ ± 2%   1.042µ ± 0%        ~ (p=0.284 n=6)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|e-12   1.994µ ± 2%   2.438µ ± 1%  +22.27% (p=0.002 n=6)
FastRegexMatcher/(?i:(zQPbMkNO|NNSPdvMi|iWuuSoAl|-12   1.998µ ± 5%   2.457µ ± 1%  +22.95% (p=0.002 n=6)
FastRegexMatcher/(?i:(zQPbMkNO.*|NNSPdvMi.*|iWuuS-12   10.06µ ± 2%   10.16µ ± 7%        ~ (p=0.240 n=6)
FastRegexMatcher/(?i:(.*zQPbMkNO|.*NNSPdvMi|.*iWu-12   12.02µ ± 2%   11.70µ ± 4%   -2.64% (p=0.041 n=6)
```

I'm experimenting with an optimization...