@pracucci 2024-05-31
> I'm experimenting with an optimization...

I think we can skip the normalization if the string only has ascii chars. The check for ascii chars is already done by `strings.ToLower()`, so if we hold our nose and copy-paste `strings.ToLower()` implementation, we can merge it with the normalization:
https://github.com/prometheus/prometheus/commit/55dc273771f409dfa98bc8bbf4ef0e1cfbcd332a

This reverts the performance impact for the common case of a string only having ASCII chars:

```
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                                     │ before.txt  │        after-optimized.txt        │
                                                     │   sec/op    │   sec/op     vs base              │
FastRegexMatcher/(?i:foo)-12                           115.5n ± 2%   112.3n ± 1%  -2.77% (p=0.009 n=6)
FastRegexMatcher/(?i:(foo|bar))-12                     239.8n ± 4%   248.6n ± 3%  +3.65% (p=0.037 n=6)
FastRegexMatcher/(?i:(foo1|foo2|bar))-12               431.2n ± 3%   437.8n ± 2%  +1.53% (p=0.039 n=6)
FastRegexMatcher/^(?i:foo|oo)|(bar)$-12                1.050µ ± 2%   1.073µ ± 2%  +2.14% (p=0.026 n=6)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|e-12   1.994µ ± 2%   1.936µ ± 1%  -2.88% (p=0.002 n=6)
FastRegexMatcher/(?i:(zQPbMkNO|NNSPdvMi|iWuuSoAl|-12   1.998µ ± 5%   1.963µ ± 4%  -1.75% (p=0.041 n=6)
FastRegexMatcher/(?i:(zQPbMkNO.*|NNSPdvMi.*|iWuuS-12   10.06µ ± 2%   10.04µ ± 3%       ~ (p=0.699 n=6)
FastRegexMatcher/(?i:(.*zQPbMkNO|.*NNSPdvMi|.*iWu-12   12.02µ ± 2%   12.15µ ± 4%       ~ (p=0.310 n=6)
geomean                                                1.252µ        1.253µ       +0.08%
```