@Ranveer777 2024-05-31
@pracucci 
Interesting approach. This will allow us to restrict the performance impact to only non-Ascii characters.