@Ranveer777 2024-06-01
@colega @bboreham 
I wanted to get your thoughts on the @pracucci approach.
Also, what are the next steps?