@colega 2024-06-01
> I wanted to get your thoughts on the @pracucci approach.

Hey, I agree that we should optimize for the ascii case, which is the most common one in the wild.

> Also, what are the next steps?

I think we can start with a middle ground between copying the entire `strings.ToLower` and just defining:

```go
func toLowerNormalized(s string) s {
    if allAscii(s) {
        return strings.ToLower()
    }
    return strings.ToLower(norm.NFKD.String(s))
}
```

Use it and benchmark to see if it's worth copying the entire `strings.ToLower()` as @pracucci suggests.