@Ranveer777 2024-06-01
@colega 
The `strings.ToLower()` implementation has already incorporated a check for ASCII characters.
```
// ToLower returns s with all Unicode letters mapped to their lower case.
func ToLower(s string) string {
	isASCII, hasUpper := true, false
	for i := 0; i < len(s); i++ {
		c := s[i]
		if c >= utf8.RuneSelf {
			isASCII = false
			break
		}
		hasUpper = hasUpper || ('A' <= c && c <= 'Z')
	}
        ...
```
Therefore, rather than employing an `if allAscii(s)` check, adopting the implementation of `strings.ToLower()`(a suggestion made by @pracucci) appears to be a more efficient approach. 

There remains a question about whether it's feasible to combine normalization and case conversion into a single, more optimized loop. What are your thoughts on this? @colega @pracucci @bboreham