@Ranveer777 2024-06-02
I've made a modification based on @pracucci suggestions. Instead of iterating through the string twice to check for uppercase characters and then converting them to lowercase for ASCII strings, we can now convert all uppercase characters to lowercase regardless of whether the string is ASCII or not.

I conducted a benchmark test, and the results were quite intriguing. I kindly request @pracucci to rerun the benchmark for re-validation.

```
                                                    │  before.txt  │              after.txt              │
                                                    │    sec/op    │    sec/op     vs base               │
FastRegexMatcher/#00-8                                74.31n ±  6%   74.14n ±  1%        ~ (p=0.699 n=6)
FastRegexMatcher/foo-8                                90.84n ±  0%   90.97n ±  2%        ~ (p=0.310 n=6)
FastRegexMatcher/^foo-8                               68.90n ±  1%   68.86n ±  1%        ~ (p=0.699 n=6)
FastRegexMatcher/(foo|bar)-8                          80.45n ±  1%   83.80n ±  4%   +4.16% (p=0.015 n=6)
FastRegexMatcher/foo.*-8                              166.0n ±  1%   165.7n ±  0%        ~ (p=0.177 n=6)
FastRegexMatcher/.*foo-8                              188.0n ±  1%   187.9n ±  0%        ~ (p=0.459 n=6)
FastRegexMatcher/^.*foo$-8                            187.9n ±  0%   187.8n ±  1%        ~ (p=0.550 n=6)
FastRegexMatcher/^.+foo$-8                            187.7n ±  2%   187.8n ±  0%        ~ (p=0.937 n=6)
FastRegexMatcher/.?-8                                 123.1n ±  0%   123.2n ±  5%        ~ (p=0.619 n=6)
FastRegexMatcher/.*-8                                 158.5n ±  1%   157.9n ±  0%        ~ (p=0.197 n=6)
FastRegexMatcher/.+-8                                 154.3n ±  0%   154.6n ±  1%        ~ (p=0.139 n=6)
FastRegexMatcher/foo.+-8                              165.1n ±  1%   164.7n ±  0%        ~ (p=0.082 n=6)
FastRegexMatcher/.+foo-8                              188.0n ±  1%   187.7n ±  0%        ~ (p=0.465 n=6)
FastRegexMatcher/foo_.+-8                             153.2n ± 13%   150.0n ±  0%        ~ (p=0.584 n=6)
FastRegexMatcher/foo_.*-8                             150.5n ±  8%   150.0n ±  0%        ~ (p=0.056 n=6)
FastRegexMatcher/.*foo.*-8                            342.9n ±  1%   343.7n ±  1%        ~ (p=0.485 n=6)
FastRegexMatcher/.+foo.+-8                            352.7n ±  2%   354.4n ±  2%        ~ (p=0.290 n=6)
FastRegexMatcher/(?s:.*)-8                            74.06n ±  1%   74.03n ±  0%        ~ (p=1.000 n=6)
FastRegexMatcher/(?s:.+)-8                            86.90n ±  0%   86.98n ±  0%   +0.10% (p=0.026 n=6)
FastRegexMatcher/(?s:^.*foo$)-8                       183.8n ±  1%   183.5n ±  0%        ~ (p=0.271 n=6)
FastRegexMatcher/(?i:foo)-8                           154.2n ±  3%   155.9n ±  4%   +1.07% (p=0.032 n=6)
FastRegexMatcher/(?i:(foo|bar))-8                     301.5n ±  2%   299.4n ±  1%        ~ (p=0.310 n=6)
FastRegexMatcher/(?i:(foo1|foo2|bar))-8               519.8n ±  2%   513.4n ±  2%        ~ (p=0.132 n=6)
FastRegexMatcher/^(?i:foo|oo)|(bar)$-8                1.270µ ±  1%   1.276µ ±  1%   +0.51% (p=0.039 n=6)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|e-8   2.630µ ±  3%   2.342µ ±  3%  -10.97% (p=0.002 n=6)
FastRegexMatcher/((.*)(bar|b|buzz)(.+)|foo)$-8        778.2n ±  1%   777.2n ±  2%        ~ (p=0.937 n=6)
FastRegexMatcher/^$-8                                 73.87n ±  1%   74.06n ±  1%        ~ (p=0.180 n=6)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-8    275.8n ±  1%   275.1n ±  1%        ~ (p=0.937 n=6)
FastRegexMatcher/10\.0\.(1|2)\.+-8                    151.4n ±  8%   150.6n ±  7%        ~ (p=0.699 n=6)
FastRegexMatcher/10\.0\.(1|2).+-8                     149.9n ±  0%   149.8n ±  1%        ~ (p=0.970 n=6)
FastRegexMatcher/((fo(bar))|.+foo)-8                  344.9n ±  1%   344.9n ±  1%        ~ (p=0.729 n=6)
FastRegexMatcher/zQPbMkNO|NNSPdvMi|iWuuSoAl|qbvKM-8   374.2n ± 19%   389.0n ±  7%        ~ (p=0.310 n=6)
FastRegexMatcher/jyyfj00j0061|jyyfj00j0062|jyyfj9-8   344.6n ± 14%   388.1n ± 10%  +12.61% (p=0.026 n=6)
FastRegexMatcher/(?i:(zQPbMkNO|NNSPdvMi|iWuuSoAl|-8   2.615µ ±  3%   2.404µ ±  3%   -8.05% (p=0.002 n=6)
FastRegexMatcher/(?i:(AAAAAAAAAAAAAAAAAAAAAAAA|BB-8   558.8n ±  1%   563.9n ±  1%   +0.92% (p=0.041 n=6)
FastRegexMatcher/(?i:(zQPbMkNO.*|NNSPdvMi.*|iWuuS-8   12.03µ ±  0%   12.05µ ±  1%        ~ (p=0.310 n=6)
FastRegexMatcher/(?i:(.*zQPbMkNO|.*NNSPdvMi|.*iWu-8   14.34µ ±  1%   14.46µ ±  1%   +0.85% (p=0.004 n=6)
FastRegexMatcher/fo.?-8                               163.0n ±  0%   162.8n ±  0%        ~ (p=0.617 n=6)
FastRegexMatcher/foo.?-8                              163.0n ±  2%   162.8n ±  0%        ~ (p=0.729 n=6)
FastRegexMatcher/f.?o-8                               158.5n ±  0%   158.5n ±  0%        ~ (p=0.623 n=6)
FastRegexMatcher/.*foo.?-8                            356.4n ±  1%   355.7n ±  1%        ~ (p=0.221 n=6)
FastRegexMatcher/.?foo.+-8                            339.6n ±  2%   340.7n ±  1%        ~ (p=0.556 n=6)
FastRegexMatcher/foo.?|bar-8                          263.2n ±  1%   262.4n ±  1%        ~ (p=0.260 n=6)
geomean                                               272.8n         272.7n         -0.06%
```

@colega @bboreham @pracucci Could I kindly ask you to spare some time for another review?